auction is open column working, credit updates working , math calculations
need to work on when auction is closed.

+++ src/main/java/com/dev/utils/Constants.java
@@ -1,7 +1,7 @@
 package com.dev.utils;
 
 public class Constants {
-public static final int STARTING_OFFERING_NUMBER = 0;
+public static final double STARTING_OFFERING_NUMBER = 0;
 public static final int MINIMAL_OFFERING_NUMBER = 1;
 public static final int OFFER_UPLOAD_COST = 1;
 public static final int PRODUCT_UPLOAD_COST = 2;
